Moved/renamespaced several classes. Made diagnostic codes ints not enums
