+++ Traction.CompileTimeDiagnosticsTests/PartialMemberTests.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
-using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using NUnit.Framework;
+using static Traction.Contracts.Rewriting.DiagnosticCodes;
+using static Traction.Roslyn.Rewriting.DiagnosticCodes;
 
 namespace Traction.DiagnosticsTests {
 
@@ -16,10 +16,10 @@ namespace Traction.DiagnosticsTests {
 
             //Assert
             if (isValid) {
-                Assert.IsFalse(diagnostics.ContainsOnlyCode(DiagnosticCode.RewriteConfirmed));
+                Assert.IsFalse(diagnostics.ContainsOnlyCode(RewriteConfirmed));
             }
             else {
-                Assert.IsTrue(diagnostics.ContainsCode(DiagnosticCode.NoContractsOnPartialMembers));
+                Assert.IsTrue(diagnostics.ContainsCode(NoContractsOnPartialMembers));
             }
         }
