+++ Traction.PlatformExtensions/Traction.PlatformExtensions.csproj
@@ -74,7 +74,7 @@
     <Compile Include="Roslyn\Syntax\PropertySyntaxExtensions.cs" />
     <Compile Include="Roslyn\Syntax\SyntaxNodeExtensions.cs" />
     <Compile Include="Roslyn\Syntax\SyntaxTokenExtensions.cs" />
-    <Compile Include="Roslyn\TypeExtensions.cs" />
+    <Compile Include="Roslyn\Reflection\TypeExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
