+++ Traction/Contracts/Rewriting/DiagnosticCodes.cs
@@ -1,4 +1,4 @@
-﻿namespace Traction {
+﻿namespace Traction.Contracts.Rewriting {
 
     /// <summary>
     /// These extend the set of codes in Traction.Roslyn.Rewriting.DiagnosticCodes,
