+++ Traction/Contracts/Rewriting/SyntaxExpanders/AutoPropertyExpander.cs
@@ -6,8 +6,9 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Traction.Roslyn.Semantics;
 using Traction.Roslyn.Syntax;
 using Traction.SEPrecompilation;
+using Traction.Contracts.Semantics;
 
-namespace Traction {
+namespace Traction.Contracts.Rewriting {
 
     /// <summary>
     /// Rewrites automatically implemented properties as normal properties with backing fields.
