+++ Traction/Contracts/Semantics/AttributeDataExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
+using Traction.Roslyn.Semantics;
 
-namespace Traction.Roslyn.Semantics {
+namespace Traction.Contracts.Semantics {
 
     internal static class AttributeDataExtensions {
