+++ Traction/Contracts/Semantics/ITypeSymbolExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using System.Linq;
 using Microsoft.CodeAnalysis;
-using Traction.Roslyn;
+using Traction.Roslyn.Semantics;
 
-namespace Traction.Roslyn.Semantics {
+namespace Traction.Contracts.Semantics {
 
     internal static class ITypeSymbolExtensions {
