+++ Traction/Contracts/Semantics/SymbolExtensionMethodsImpl.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.CodeAnalysis;
-using Traction.Roslyn;
+using Traction.Roslyn.Semantics;
 
-namespace Traction.Roslyn.Semantics {
+namespace Traction.Contracts.Semantics {
 
     internal static class SymbolExtensionMethodsImpl {
