+++ Traction/Traction.csproj
@@ -64,36 +64,36 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Roslyn\Semantics\AttributeDataExtensions.cs" />
+    <Compile Include="Contracts\Semantics\AttributeDataExtensions.cs" />
     <Compile Include="ClientAPI\ContractAttribute.cs" />
     <Compile Include="ClientAPI\ContractAttributes.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>ContractAttributes.tt</DependentUpon>
     </Compile>
-    <Compile Include="Roslyn\Syntax\AttributeSyntaxExtenions.cs" />
-    <Compile Include="Roslyn\Semantics\ITypeSymbolExtensions.cs" />
-    <Compile Include="Roslyn\Semantics\SymbolExtensionMethodsImpl.cs" />
-    <Compile Include="ContractProvider.cs" />
-    <Compile Include="Contract.cs" />
-    <Compile Include="Roslyn\Semantics\IParameterSymbolExtensions.cs" />
-    <Compile Include="Roslyn\Semantics\IPropertySymbolExtensions.cs" />
-    <Compile Include="IContractProvider.cs" />
-    <Compile Include="Roslyn\Semantics\IMethodSymbolExtensions.cs" />
-    <Compile Include="Rewriters\DiagnosticChecker.cs" />
-    <Compile Include="ClientAPI\DiagnosticCode.cs" />
+    <Compile Include="Contracts\Syntax\AttributeSyntaxExtenions.cs" />
+    <Compile Include="Contracts\Semantics\ITypeSymbolExtensions.cs" />
+    <Compile Include="Contracts\Semantics\SymbolExtensionMethodsImpl.cs" />
+    <Compile Include="Contracts\Semantics\ContractProvider.cs" />
+    <Compile Include="Contracts\Semantics\Contract.cs" />
+    <Compile Include="Contracts\Semantics\IParameterSymbolExtensions.cs" />
+    <Compile Include="Contracts\Semantics\IPropertySymbolExtensions.cs" />
+    <Compile Include="Contracts\Semantics\IContractProvider.cs" />
+    <Compile Include="Contracts\Semantics\IMethodSymbolExtensions.cs" />
+    <Compile Include="Contracts\Rewriting\DiagnosticChecker.cs" />
+    <Compile Include="Contracts\Rewriting\DiagnosticCodes.cs" />
     <Compile Include="ClientAPI\PreconditionException.cs" />
     <Compile Include="ClientAPI\PostconditionException.cs" />
-    <Compile Include="Rewriters\SyntaxExpanders\AutoPropertyExpander.cs" />
-    <Compile Include="Rewriters\SyntaxExpanders\IteratorBlockExpander.cs" />
-    <Compile Include="Rewriters\SyntaxExpanders\ExpressionBodiedMemberExpander.cs" />
-    <Compile Include="Rewriters\IdentifierFactory.cs" />
+    <Compile Include="Contracts\Rewriting\SyntaxExpanders\AutoPropertyExpander.cs" />
+    <Compile Include="Contracts\Rewriting\SyntaxExpanders\IteratorBlockExpander.cs" />
+    <Compile Include="Contracts\Rewriting\SyntaxExpanders\ExpressionBodiedMemberExpander.cs" />
+    <Compile Include="Contracts\Rewriting\IdentifierFactory.cs" />
     <Compile Include="TractionCompileModule.cs" />
-    <Compile Include="ClientAPI\DiagnosticFactory.cs" />
-    <Compile Include="Rewriters\ContractRewriter.cs" />
+    <Compile Include="Contracts\Rewriting\DiagnosticFactory.cs" />
+    <Compile Include="Contracts\Rewriting\ContractRewriter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Rewriters\RewriterBase.cs" />
-    <Compile Include="Rewriters\ConcreteTypeRewriter.cs" />
+    <Compile Include="Contracts\Rewriting\RewriterBase.cs" />
+    <Compile Include="Contracts\Rewriting\ConcreteTypeRewriter.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
